--- a/.github/workflows/release-pypi.yml
+++ b/.github/workflows/release-pypi.yml
@@ -21,11 +21,8 @@ jobs:
         with:
           python-version: 3.9
 
-      - name: Install pypa/build
-        run: python -m pip install build
-
       - name: Build a binary wheel and a source tarball
-        run: python -m build --sdist --wheel --outdir dist/
+        run: make build
 
       - name: Release on PyPI
         uses: pypa/gh-action-pypi-publish@master