--- a/httpie/internal/_build_channel.py
+++ b/httpie/internal/_build_channel.py
@@ -1,0 +1,5 @@
+# Represents the packaging method. This file should
+# be overridden by every build system we support on
+# the packaging step.
+
+BUILD_CHANNEL = 'unknown'