… docs/README.md
… httpie/internal/__build_channel__.py
… httpie/utils.py
… tests/utils/__init__.py